some bug fixing
Former-commit-id: 442a0a311408fad16cb780fe92cbb5b9bbf9443a
Former-commit-id: 9eb09ac12bcfc04ae71614f7753be9e0fa4bcc92

+++ aws-upload/scripts/stack.py
@@ -26,7 +26,7 @@ class Stack:
 		#When we use a large stack, there will be some more complex logic here to retrieve the chunck
 		#because we will load the required z-planes as needed.
 
-		return self.input[z_min:z_max, y_min:y_max, x_min:x_max]
+		return self.input[z_min:z_max, y_min:y_max, x_min:x_max].astype('double')
 
 	def convertToHDF5(self, fov = numpy.array([8, 172 ,172])):
